requirements/production.txt
django-environ==0.11.2
cloudinary==1.36.0
dj3-cloudinary-storage==0.0.6
whitenoise==6.5.0